+++ +sbfsem/+io/OData.m
@@ -203,11 +203,11 @@ classdef OData < handle
             % Optional input:
             %   numStructures       Number to return (default = 1)
             % -------------------------------------------------------------
-            str = [getServiceRoot(obj.source), 'Structures(',...
-                   num2str(ID), ')$filter=Username eq ', username,...
-                   '&$select=ID,ParentID&$orderby=LastModified desc'];
+            str = [getServiceRoot(obj.source), 'Structures?',...
+                   '$filter=Username eq ''', username,...
+                   '''&$select=ID,ParentID&$orderby=LastModified desc'];
             if nargin == 3
-                str = [str, '&$top=', numStructures];
+                str = [str, '&$top=', num2str(numStructures)];
             end
             data = webread(str, obj.webOpt);
             data = cat(1, data.value{:});